How to write to the SD Card on the Adalogger
The SDCardInfo sketch is basically a copy of the Example sketch for SD -> CardInfo, to report on the stats of your SD card (if a card is, indeed, in the slot).

The Adalogger sketch uses a TMP36 analog sensor, and the i2c Real-Time Clock, putting a formatted time string, and temperature reading, and the LiPo battery voltage into the SD card records, and also printing to the Serial Monitor. The data is logged every 10 seconds. and the data sets are delimited by semi-colons, and each set is delimited by commas. It looks like this;

2017-11-4 22:18:34, PT, Saturday; TempC, 22.51; TempF, 72.52; Batt, 4.26