Small tuning updates.
The Adalogger sketch seems to open the file successfully, but doesn't seem to be putting any data INTO the file(s). Every time I upload a new change to the sketch, it's incrementing the filename by one... I also want to use the RTC to have it create the file with a useful timestamp!